requirements.txt
alembic==1.14.0
anyio==4.8.0
certifi==2024.12.14
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
python-dotenv==1.0.1
python-telegram-bot==21.10
sniffio==1.3.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
